GEODE-6096: Move MethExecutor classes into dunit standalone
