+++ geode-dunit/src/main/java/org/apache/geode/test/dunit/standalone/RemoteDUnitVM.java
@@ -18,8 +18,6 @@ import java.rmi.RemoteException;
 import java.rmi.server.UnicastRemoteObject;
 import java.util.concurrent.TimeUnit;
 
-import hydra.MethExecutor;
-import hydra.MethExecutorResult;
 import org.apache.logging.log4j.Logger;
 
 import org.apache.geode.internal.logging.LogService;
